# html/docs/openapi.yaml
openapi: 3.1.1
info:
  title: API Médicaments FR
  description: |
    API haute performance pour accéder aux données des médicaments français issus de la Base de Données Publique des Médicaments (BDPM).

    **Caractéristiques principales :**
    - Données officielles mises à jour automatiquement 2 fois par jour (6h et 18h)
    - Accès gratuit avec rate limiting équitable (1000 tokens initiaux, recharge à 3 tokens/sec)
    - Données intègres et non altérées depuis la source BDPM
    - Endpoints v1 modernes + endpoints legacy (dépréciés, suppression prévue le 31 juillet 2026)

    **Remarque :** Les endpoints legacy sont toujours fonctionnels mais seront supprimés. Veuillez migrer vers les endpoints v1.

    **Ressources utiles:**
    - [Guide de migration vers v1](https://github.com/Giygas/medicaments-api/blob/main/docs/MIGRATION.md)
    - [Base de données BDPM](https://base-donnees-publique.medicaments.gouv.fr)

  version: 1.1.0
  contact:
    name: GitHub Repository
    url: https://github.com/giygas/medicaments-api
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: https://medicaments-api.giygas.dev/
security: []
...
paths:
...
  /v1/medicaments/export:
    get:
      summary: Exporter tous les médicaments (v1)
      description: |
        Exporter la base de données complète de tous les médicaments.
        Utilisez ETag et Last-Modified headers pour le cache conditionnel.
      tags:
        - Médicaments (v1)
      responses:
        "200":
          description: Réponse réussie avec liste complète
          headers:
            ETag:
              schema:
                type: string
              examples:
                etag-header:
                  value: 'W/"abc123"'
            Last-Modified:
              schema:
                type: string
                format: date-time
              examples:
                last-modified-header:
                  value: "Mon, 15 Jan 2026 06:00:00 GMT"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicamentArray"
              examples:
                export:
                  summary: Export complet
                  value:
                    - cis: "61504672"
                      elementPharmaceutique: "PARACETAMOL MYLAN 1 g, comprimé"
                      formePharmaceutique: "comprimé"
                      voiesAdministration: ["orale"]
                      statusAutorisation: "Autorisation active"
                      typeProcedure: "Procédure nationale"
                      etatComercialisation: "Commercialisée"
                      dateAMM: "2000-01-01"
                      titulaire: "MYLAN SAS"
                      surveillanceRenforcee: "Non"
                      composition: []
                      generiques: []
                      presentation: []
                      conditions: []
                    - cis: "61504672"
                      elementPharmaceutique: "PARACETAMOL 500 mg, comprimé"
                      formePharmaceutique: "comprimé"
                      voiesAdministration: ["orale"]
                      statusAutorisation: "Autorisation active"
                      typeProcedure: "Procédure nationale"
                      etatComercialisation: "Commercialisée"
                      dateAMM: "2000-01-01"
                      titulaire: "MYLAN SAS"
                      surveillanceRenforcee: "Non"
                      composition: []
                      generiques: []
                      presentation: []
                      conditions: []
        "400":
          description: Code CIS invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                bad-request:
                  value:
                    error: "Bad Request"
                    message: "Invalid CIS"
                    code: 400
        "404":
          description: Médicament non trouvé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                not-found:
                  value:
                    error: "Not Found"
                    message: "Medicament not found"
                    code: 404
        "500":
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                internal-error:
                  value:
                    error: "Internal Server Error"
                    message: "Internal server error"
                    code: 500
...
  /database:
    get:
      deprecated: true
      summary: "[LEGACY] Obtenir tous les médicaments"
      description: |
        **DEPRÉCIÉ - Ce endpoint sera supprimé le 31 juillet 2026**

        Récupérer la liste complète de tous les médicaments dans la base de données.

        **Migration :** Utilisez `/v1/medicaments/export` à la place.
      tags:
        - Médicaments (Legacy)
      responses:
        "200":
          description: Réponse réussie
          headers:
            Deprecation:
              schema:
                type: string
              examples:
                deprecation-header:
                  value: "true"
            Sunset:
              schema:
                type: string
                format: date-time
              examples:
                sunset-date:
                  value: "2026-07-31T23:59:59Z"
            Link:
              schema:
                type: string
              examples:
                successor-version:
                  value: '<https://medicaments-api.giygas.dev/v1/medicaments/export>; rel="successor-version"'
            X-Deprecated:
              schema:
                type: string
              examples:
                x-deprecated-header:
                  value: "Use /v1/medicaments/export instead"
            Warning:
              schema:
                type: string
              examples:
                warning-header:
                  value: '299 - "Deprecated endpoint"'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medicament"
              examples:
                database:
                  summary: Liste complète (legacy)
                  value:
                    - cis: "61504672"
                      elementPharmaceutique: "PARACETAMOL MYLAN 1 g, comprimé"
                      formePharmaceutique: "comprimé"
                      voiesAdministration: ["orale"]
                      statusAutorisation: "Autorisation active"
                      typeProcedure: "Procédure nationale"
                      etatComercialisation: "Commercialisée"
                      dateAMM: "2000-01-01"
                      titulaire: "MYLAN SAS"
                      surveillanceRenforcee: "Non"
                      composition: []
                      generiques: []
                      presentation: []
                      conditions: []
                    - cis: "61504672"
                      elementPharmaceutique: "PARACETAMOL 500 mg, comprimé"
                      formePharmaceutique: "comprimé"
                      voiesAdministration: ["orale"]
                      statusAutorisation: "Autorisation active"
                      typeProcedure: "Procédure nationale"
                      etatComercialisation: "Commercialisée"
                      dateAMM: "2000-01-01"
                      titulaire: "MYLAN SAS"
                      surveillanceRenforcee: "Non"
                      composition: []
                      generiques: []
                      presentation: []
                      conditions: []
        "500":
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    Medicament:
      type: object
      title: Medicament
      properties:
        cis:
          type: string
          pattern: "^[0-9]{1,9}$"
          title: Code CIS
        elementPharmaceutique:
          type: string
          title: Nom pharmaceutique
        formePharmaceutique:
          type: string
          title: Forme pharmaceutique
        voiesAdministration:
          type: array
          items:
            type: string
          title: Voies d'administration
        statusAutorisation:
          type: string
          title: Statut d'autorisation
        typeProcedure:
          type: string
          title: Type de procédure
        etatComercialisation:
          type: string
          title: Statut de commercialisation
        dateAMM:
          type: string
          format: date
          title: Date AMM
        titulaire:
          type: string
          title: Titulaire de l'autorisation
        surveillanceRenforcee:
          type: string
          title: Surveillance renforcée
        composition:
          type: array
          items:
            $ref: "#/components/schemas/CompositionItem"
          title: Composition
        generiques:
          type: array
          items:
            $ref: "#/components/schemas/GeneriqueItem"
          title: Génériques associés
        presentation:
          type: array
          items:
            $ref: "#/components/schemas/PresentationResponse"
          title: Présentations disponibles
        conditions:
          type: array
          items:
            type: string
          title: Conditions de prescription
...
    PresentationResponse:
      type: object
      title: PresentationResponse
      properties:
        cis:
          type: string
          pattern: "^[0-9]{1,9}$"
          title: Code CIS
        cip7:
          type: string
          pattern: "^[0-9]{7}$"
          title: Code CIP-7
        libelle:
          type: string
          title: Libellé de présentation
        statusAdministratif:
          type: string
          title: Statut administratif
        etatComercialisation:
          type: string
          title: Statut de commercialisation
        dateDeclaration:
          type: string
          format: date
          title: Date de déclaration
        cip13:
          type: string
          pattern: "^[0-9]{13}$"
          title: Code CIP-13
        agreement:
          type: string
          title: Statut d'agrément
        tauxRemboursement:
          type: string
          title: Taux de remboursement
        prix:
          type: number
          format: float
          title: Prix
...
    ErrorResponse:
      type: object
      title: ErrorResponse
      properties:
        error:
          type: string
          title: Statut d'erreur
        message:
          type: string
          title: Message d'erreur
        code:
          type: integer
          title: Code de statut HTTP
    MedicamentArray:
      type: array
      title: MedicamentArray
      items:
        $ref: "#/components/schemas/Medicament"
    CompositionItem:
      type: object
      title: CompositionItem
      properties:
        cis:
          type: string
          pattern: "^[0-9]{1,9}$"
          title: Code CIS
        elementPharmaceutique:
          type: string
          title: Élément pharmaceutique
        codeSubstance:
          type: string
          pattern: "^[0-9]+$"
          title: Code de substance
        denominationSubstance:
          type: string
          title: Dénomination de substance
        dosage:
          type: string
          title: Dosage
        referenceDosage:
          type: string
          title: Dosage de référence
        natureComposant:
          type: string
          title: Nature du composant
    GeneriqueItem:
      type: object
      title: GeneriqueItem
      properties:
        cis:
          type: string
          pattern: "^[0-9]{1,9}$"
          title: Code CIS
        group:
          type: string
          pattern: "^[0-9]+$"
          title: ID de groupe générique
        libelle:
          type: string
          title: Libellé de groupe générique
...